+++ tests/tbenchmarks2.nim
@@ -25,7 +25,7 @@ Original benchmarks results:
 type
   Emitter* = ref object of Agent
   Counter* = ref object of Agent
-    value: int
+    value: int64
 
 proc bump*(tp: Emitter, val: array[1024, int]) {.signal.}
 
@@ -45,7 +45,7 @@ var
 const n = block:
   when defined(slowbench): 1_000_000
   else: 100_000
-const expectedValue = (n * (n - 1)) div 2
+const expectedValue = (int64(n) * (n - 1)) div 2
 
 suite "benchmarks":
   test "slot proc baseline (tight loop)":
